third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-svgdom-k2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k2' propoerty of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(overComposite1.k2.baseVal * 1000) is 900
PASS Math.round(overComposite1.k2.baseVal * 1000) is 100
PASS successfullyParsed is true

TEST COMPLETE
